requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
streamlit>=1.40.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
beautifulsoup4>=4.12.0
requests>=2.32.0
pandas>=2.2.0
pydantic>=2.0.0
aiofiles>=24.0.0
httpx>=0.27.0
pypdf>=4.0.0
sse-starlette>=2.1.0
langchain-groq>=0.2.0
psutil>=5.9.0
fpdf2>=2.8.0
